Adds getAppFileprotection logic to Apps client
